Support cellular RSSI on IOS XE (#10726)
Move Cisco cellular RSSI to a trait and add it to IOS XE
