+++ LibreNMS/OS/Traits/CiscoCellular.php
@@ -1,0 +1,57 @@
+<?php
+/**
+ * CiscoCellular.php
+ *
+ * -Description-
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @package    LibreNMS
+ * @link       http://librenms.org
+ * @copyright  2019 Tony Murray
+ * @author     Tony Murray <murraytony@gmail.com>
+ */
+
+namespace LibreNMS\OS\Traits;
+
+use LibreNMS\Device\WirelessSensor;
+
+trait CiscoCellular
+{
+    /**
+     * Discover wireless RSSI (Received Signal Strength Indicator). This is in dBm. Type is rssi.
+     * Returns an array of LibreNMS\Device\Sensor objects that have been discovered
+     *
+     * @return array
+     */
+    public function discoverWirelessRssi()
+    {
+        $sensors = array();
+
+        $data = snmpwalk_cache_oid($this->getDevice(), 'c3gCurrentGsmRssi', array(), 'CISCO-WAN-3G-MIB');
+        foreach ($data as $index => $entry) {
+            $sensors[] = new WirelessSensor(
+                'rssi',
+                $this->getDeviceId(),
+                '.1.3.6.1.4.1.9.9.661.1.3.4.1.1.1.' . $index,
+                'ios',
+                $index,
+                'RSSI: Chain ' . str_replace('1.', '', $index),
+                $entry['c3gCurrentGsmRssi.1']
+            );
+        }
+
+        return $sensors;
+    }
+}
